feat(as-5031): set up integration testing for document-service-api
