+++ packages/document-service-api/src/routes/index.js
@@ -14,6 +14,5 @@ const router = Router();
 router.use('/api-docs', apiDocs);
 router.use('/api/v1/migrate-metadata', migrateMetadata);
 router.use('/api/v1/:applicationId', application);
-router.use('/test', () => {body: "hey"});
 
 module.exports = router;
