+++ packages/document-service-api/test/developer/globalSetup.ts
@@ -1,9 +1,4 @@
 module.exports = async () => {
-	const {
-		createMongoContainer,
-		destroyMongoContainer,
-		containerNotSpunUp
-	} = require('./mongodb-container-helper');
-	
+	const { createMongoContainer } = require('./mongodb-container-helper');
 	await createMongoContainer();
 };
