+++ packages/document-service-api/test/developer/globalTeardown.ts
@@ -1,0 +1,4 @@
+module.exports = async () => {
+	const { destroyMongoContainer } = require('./mongodb-container-helper');
+	await destroyMongoContainer();
+};
